gdalcompare.py: compute difference on float to avoid integer underflow
git-svn-id: https://svn.osgeo.org/gdal/trunk@27795 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/swig/python/scripts/gdalcompare.py
@@ -69,7 +69,7 @@ def compare_image_pixels(golden_band, new_band, id, options=[]):
   for line in range(golden_band.YSize):
     golden_line = golden_band.ReadAsArray(0, line, golden_band.XSize, 1)[0]
     new_line = new_band.ReadAsArray(0, line, golden_band.XSize, 1)[0]
-    diff_line = golden_line - new_line
+    diff_line = golden_line.astype(float) - new_line.astype(float)
     max_diff = max(max_diff,abs(diff_line).max())
     diff_count += len(diff_line.nonzero()[0])
